Barbershop Hook-in, Restore Previous Character in Character Menu
