--- resources/orp/client/html/barbershop/style.less
@@ -1,221 +1,0 @@
-@sidebarWidth: 320px;
-@tabHeight: 50px;
-@footerHeight: 25px;
-
-@dark: #141414e0;
-@light: #d7e1e5;
-@contrast1: rgb(0, 217, 255);
-@contrast2: #7d909b;
-@darkaccent: rgb(8, 8, 8);
-@fontface: 'Roboto';
-
-body,
-html {
-    height: 100%;
-    margin: 0;
-    padding: 0;
-    background: transparent;
-    font-family: @fontface;
-    text-transform: uppercase;
-    user-select: none;
-}
-
-.data-menu {
-    width: @sidebarWidth;
-    height: 100%;
-    float: left;
-    color: white;
-}
-
-#app {
-    display: grid;
-    grid-template-columns: @sidebarWidth;
-    grid-template-rows: @tabHeight 1fr;
-    height: 100%;
-}
-
-.tab {
-    width: 97.5%;
-    background: @dark;
-    text-align: center;
-    line-height: 30px;
-    box-shadow: 8px 0px 0px 0px @darkaccent;
-
-    .title {
-        font-size: 16px;
-        font-family: @fontface;
-        font-weight: 100;
-        color: @contrast1;
-        transition: 0.1s;
-    }
-
-    .title:hover {
-        color: white;
-        user-select: none;
-    }
-}
-
-.footer {
-    position: fixed;
-    width: 10vw;
-    height: 5vh;
-    right: 0;
-    bottom: 0;
-    margin: 25px;
-    background: @dark;
-    text-align: center;
-    font-size: 12px;
-    line-height: 5vh;
-    cursor: pointer;
-    transition: 0.2s;
-}
-
-.footer:hover {
-    background: @dark;
-    box-shadow: inset 0px -2px 0px 0px @contrast1;
-    color: @contrast1;
-}
-
-.footer:active {
-    background: @light;
-    color: @dark;
-}
-
-/*********************************/
-/*********      MODS     *********/
-/*********************************/
-
-.mod-list {
-    overflow-y: auto;
-    background: @dark;
-}
-
-.mod {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    grid-auto-rows: auto;
-    grid-row-gap: 10px;
-    grid-template-areas:
-        'title count'
-        'switcher switcher';
-    padding: 20px;
-    transition: 0.2s;
-    font-size: 10px;
-
-    .item-switcher {
-        display: none;
-        transition: 2s;
-    }
-
-    .title {
-        grid-area: title;
-        transition: 2s;
-    }
-
-    .count {
-        grid-area: count;
-        text-align: right;
-        transition: 2s;
-    }
-}
-
-.mod:hover {
-    background-color: rgba(255, 255, 255, 0.1);
-    box-shadow: inset 3px 0px 0px 0px @contrast1;
-
-    .item-switcher {
-        display: flex;
-    }
-
-    .title,
-    .count {
-        color: @contrast1;
-    }
-}
-
-/********************************/
-/****      Item Switcher     ****/
-/********************************/
-
-.item-switcher {
-    grid-area: switcher;
-    width: 100%;
-    text-align: center;
-    padding: 5px;
-
-    display: flex;
-    justify-content: space-between;
-
-    button {
-        background: rgba(0, 0, 0, 0);
-        border: 0;
-        color: #feffff;
-        font-weight: bold;
-        transition: 1s;
-    }
-
-    button:hover {
-        cursor: pointer;
-        color: @contrast1;
-    }
-}
-
-.scroll::-webkit-scrollbar {
-    width: 8px;
-    background: @darkaccent;
-    box-shadow: -1px 0px 0px 0px rgba(0, 0, 0, 0.25);
-}
-
-.scroll::-webkit-scrollbar-thumb {
-    background: @contrast1;
-    box-shadow: 0px 2px 0px 0px rgba(0, 0, 0, 0.1);
-}
-
-.scroll::-webkit-scrollbar-thumb:hover {
-    cursor: pointer !important;
-}
-
-input[type='range'] {
-    -webkit-appearance: none;
-    width: 100%;
-    background: transparent;
-}
-
-input[type='range']:focus {
-    outline: none;
-}
-
-input[type='range']::-webkit-slider-thumb {
-    -webkit-appearance: none;
-}
-
-input[type='range']::-ms-track {
-    width: 100%;
-    cursor: pointer;
-
-    /* Hides the slider so custom styles can be added */
-    background: transparent;
-    border-color: transparent;
-    color: transparent;
-}
-
-input[type='range']::-webkit-slider-thumb {
-    -webkit-appearance: none;
-    height: 20px;
-    width: 10px;
-    background: white;
-    margin-top: -7px;
-    cursor: pointer;
-    transition: 1s;
-}
-
-input[type='range']::-webkit-slider-thumb:hover {
-    background: @contrast1;
-}
-
-input[type='range']::-webkit-slider-runnable-track {
-    width: 100%;
-    height: 5px;
-    cursor: pointer;
-    background: rgba(255, 255, 255, 0.15);
-}
