--- resources/orp/client/html/character-new/style.less
@@ -1,307 +1,0 @@
-@sidebarWidth: 320px;
-@tabHeight: 50px;
-@footerHeight: 25px;
-
-@dark: #141414e0;
-@light: #d7e1e5;
-@contrast1: rgb(0, 217, 255);
-@contrast2: rgb(255, 153, 0);
-@darkaccent: rgb(8, 8, 8);
-@fontface: 'Roboto';
-
-body,
-html {
-    height: auto;
-    margin: 0;
-    padding: 0;
-    background: transparent;
-    font-family: @fontface;
-    text-transform: uppercase;
-    user-select: none;
-    font-size: 12px;
-}
-
-#render {
-    width: 100%;
-}
-
-.categories {
-    position: absolute;
-    display: flex;
-    flex-direction: column;
-    right: 0;
-    width: 20%;
-    color: white;
-    padding: 20px;
-    min-height: 100%;
-
-    .group {
-        padding-top: 10px;
-        padding-bottom: 10px;
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-
-        .group-header {
-            background: @dark;
-            height: 35px;
-            width: 100%;
-            display: flex;
-            flex-direction: row-reverse;
-            justify-content: space-between;
-            justify-items: center;
-            align-items: center;
-            align-content: center;
-            border: 2px solid rgba(255, 255, 255, 0.1);
-
-            .group-title {
-                width: 100%;
-                text-align: center;
-            }
-
-            button {
-                width: 35px;
-                height: 100%;
-                background: none;
-                color: @contrast1;
-                border: 2px solid rgba(255, 255, 255, 0.1);
-                border-top: 0px;
-                border-left: 0px;
-                border-bottom: 0px;
-            }
-
-            button:focus {
-                outline: none;
-            }
-        }
-
-        .group-header:hover {
-            border: 2px solid rgba(255, 255, 255, 0.1);
-
-            button {
-                background: @contrast2;
-                color: white;
-                border: 2px solid rgba(255, 255, 255, 0.1);
-                border-top: 0px;
-                border-left: 0px;
-                border-bottom: 0px;
-            }
-        }
-
-        .rows {
-            display: flex;
-            flex-direction: column;
-            border: 2px solid rgba(255, 255, 255, 0.1);
-            border-top: 0px;
-            width: 100%;
-            background: @dark;
-            justify-content: center;
-            justify-items: center;
-            align-content: center;
-            align-items: center;
-
-            .row {
-                display: flex;
-                flex-direction: column;
-                justify-content: center;
-                justify-items: center;
-                align-content: center;
-                align-items: center;
-                height: 75px;
-                width: 90%;
-
-                input {
-                    width: 100%;
-                }
-            }
-
-            .row-tattoo {
-                .label {
-                    width: 100%;
-                }
-
-                display: flex;
-                flex-direction: row;
-                justify-content: space-between;
-                justify-content: center;
-                align-content: center;
-                align-items: center;
-                height: 35px;
-                width: 90%;
-            }
-        }
-    }
-}
-
-.rotation {
-    left: 33%;
-    position: fixed;
-    max-width: 33%;
-    bottom: 25px;
-}
-
-.zoom {
-    left: 33%;
-    position: fixed;
-    max-width: 33%;
-    top: 25px;
-}
-
-.zpos {
-    left: 5%;
-    position: fixed;
-    min-width: 33vh;
-    max-width: 33vh;
-    transform: rotate(270deg);
-    top: 66vh;
-    transform-origin: left;
-}
-
-.savechanges {
-    position: fixed;
-    bottom: 25px;
-    left: 25px;
-    width: 200px;
-    height: 50px;
-    background: @dark;
-    border: 2px solid rgba(255, 255, 255, 0.2);
-    color: @contrast1;
-}
-
-.savechanges:hover {
-    border: 2px solid @contrast2;
-    color: @contrast2;
-}
-
-::-webkit-scrollbar {
-    width: 8px;
-    background: @darkaccent;
-    box-shadow: -1px 0px 0px 0px rgba(0, 0, 0, 0.25);
-}
-
-::-webkit-scrollbar-thumb {
-    background: rgba(255, 255, 255, 0.2);
-    box-shadow: 0px 2px 0px 0px rgba(0, 0, 0, 0.1);
-}
-
-::-webkit-scrollbar-thumb:hover {
-    cursor: pointer !important;
-}
-
-@track-color: black;
-@thumb-color: @contrast2;
-
-@thumb-radius: 0px;
-@thumb-height: 15px;
-@thumb-width: 10px;
-@thumb-shadow-size: 1px;
-@thumb-shadow-blur: 1px;
-@thumb-shadow-color: #111;
-@thumb-border-width: 0px;
-@thumb-border-color: white;
-
-@track-width: 100%;
-@track-height: 2px;
-@track-shadow-size: 0px;
-@track-shadow-blur: 0px;
-@track-shadow-color: #222;
-@track-border-width: 0px;
-@track-border-color: black;
-
-@track-radius: 0px;
-@contrast: 5%;
-
-.shadow(@shadow-size,@shadow-blur,@shadow-color) {
-    box-shadow: @shadow-size @shadow-size @shadow-blur @shadow-color,
-        0px 0px @shadow-size lighten(@shadow-color, 5%);
-}
-
-.track() {
-    width: @track-width;
-    height: @track-height;
-    cursor: pointer;
-    animate: 0.2s;
-}
-
-.thumb() {
-    .shadow(@thumb-shadow-size, @thumb-shadow-blur, @thumb-shadow-color);
-    border: @thumb-border-width solid @thumb-border-color;
-    height: @thumb-height;
-    width: @thumb-width;
-    border-radius: @thumb-radius;
-    background: @thumb-color;
-    cursor: pointer;
-}
-
-input[type='range'] {
-    -webkit-appearance: none;
-    margin: @thumb-height / 2 0;
-    width: @track-width;
-
-    &:focus {
-        outline: none;
-    }
-
-    &::-webkit-slider-runnable-track {
-        .track();
-        .shadow(@track-shadow-size, @track-shadow-blur, @track-shadow-color);
-        background: @track-color;
-        border-radius: @track-radius;
-        border: @track-border-width solid @track-border-color;
-    }
-
-    &::-webkit-slider-thumb {
-        .thumb();
-        -webkit-appearance: none;
-        margin-top: ((-@track-border-width * 2 + @track-height) / 2) - (@thumb-height / 2);
-    }
-
-    &::-webkit-slider-thumb:hover {
-        background: @contrast1;
-    }
-
-    &:focus::-webkit-slider-runnable-track {
-        background: lighten(@track-color, @contrast);
-    }
-
-    &::-moz-range-track {
-        .track();
-        .shadow(@track-shadow-size, @track-shadow-blur, @track-shadow-color);
-        background: @track-color;
-        border-radius: @track-radius;
-        border: @track-border-width solid @track-border-color;
-    }
-    &::-moz-range-thumb {
-        .thumb();
-    }
-
-    &::-ms-track {
-        .track();
-        background: transparent;
-        border-color: transparent;
-        border-width: @thumb-width 0;
-        color: transparent;
-    }
-
-    &::-ms-fill-lower {
-        background: darken(@track-color, @contrast);
-        border: @track-border-width solid @track-border-color;
-        border-radius: @track-radius*2;
-        .shadow(@track-shadow-size, @track-shadow-blur, @track-shadow-color);
-    }
-    &::-ms-fill-upper {
-        background: @track-color;
-        border: @track-border-width solid @track-border-color;
-        border-radius: @track-radius*2;
-        .shadow(@track-shadow-size, @track-shadow-blur, @track-shadow-color);
-    }
-    &::-ms-thumb {
-        .thumb();
-    }
-    &:focus::-ms-fill-lower {
-        background: @track-color;
-    }
-    &:focus::-ms-fill-upper {
-        background: lighten(@track-color, @contrast);
-    }
-}
