+++ resources/orp/server/utility/locationhelper.mjs
@@ -3,8 +3,18 @@ import * as alt from 'alt';
 export function goToLastLocation(player) {
     if (player.lastLocation === undefined) return;
 
-    alt.log(`${player.name} was teleported to their last location.`);
+    setTimeout(() => {
+        alt.emit('goto:LastLocation', player);
+    }, 250);
+}
 
+alt.on('goto:LastLocation', player => {
+    if (!player) return;
+    alt.log(`${player.name} was teleported to their last location.`);
     player.pos = player.lastLocation;
     player.lastLocation = undefined;
-}
+});
+
+alt.onClient('utility:SetLastLocation', (player, pos) => {
+    player.lastLocation = pos;
+});
